# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Productos - Básico 5",
        "description": "Documentación de la API REST para la gestión de productos, categorías, marcas y fotos",
        "contact": {
            "email": "soporte@ejemplo.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/mvc-productos/basico5/public",
            "description": "Servidor de desarrollo"
        }
    ],
    "paths": {
...
        "/api/v1/productos/{producto}/fotos": {
            "post": {
                "tags": [
                    "Fotos de Productos"
                ],
                "summary": "Subir fotos de producto",
                "description": "Sube una o más fotos para un producto específico",
                "operationId": "storeProductoFoto",
                "parameters": [
                    {
                        "name": "producto",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "fotos"
                                ],
                                "properties": {
                                    "fotos": {
                                        "description": "Archivos de imagen (máximo 10MB cada uno)",
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Fotos cargadas exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Fotos cargadas."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ProductoFoto"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    },
                    "422": {
                        "description": "Datos de entrada inválidos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/productos/{producto}/fotos/{foto}": {
            "delete": {
                "tags": [
                    "Fotos de Productos"
                ],
                "summary": "Eliminar foto de producto",
                "description": "Elimina una foto específica de un producto",
                "operationId": "destroyProductoFoto",
                "parameters": [
                    {
                        "name": "producto",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "foto",
                        "in": "path",
                        "description": "ID de la foto",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Foto eliminada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Foto eliminada."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "404": {
                        "description": "Producto o foto no encontrados"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "ProductoFoto": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "producto_id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "url": {
                        "type": "string",
                        "format": "uri",
                        "example": "http://localhost/storage/fotos/abc123.jpg"
                    },
                    "ruta": {
                        "type": "string",
                        "example": "fotos/abc123.jpg"
                    },
                    "nombre_original": {
                        "type": "string",
                        "example": "foto.jpg"
                    },
                    "mime_type": {
                        "type": "string",
                        "example": "image/jpeg"
                    },
                    "tamano_bytes": {
                        "type": "integer",
                        "example": 102400
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Ingrese el token JWT con el prefijo \"Bearer \" (ejemplo: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...)",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}